Added new graph programs; fix stdout calls to stdin; fix timer type.
 Changes to be committed:
	modified:   Makefile
	new file:   graph_programs/ball1.c
	new file:   graph_programs/ball2.c
	new file:   graph_programs/line1.c
	new file:   graph_programs/line2.c
	new file:   graph_programs/line4.c
	new file:   graph_programs/line5.c
	new file:   graph_programs/pixel.c

1. Added new graph programs.

	modified:   include/graph.h
	modified:   windows/graph.c

1. Fix timer type to long to match services.c.

	modified:   include/localdefs.h

1. gcc now requires () on define (apparently).

	modified:   terminal_games/mine.c
	modified:   terminal_games/snake.c
	modified:   terminal_programs/wator.c

1. Fix stdout on timer call to stdin. The manual specifies this call must
be against the window file.

Note:

Several example programs for graph were translated and run more or less
correctly. Lots of corner case errors.

+++ include/graph.h
@@ -320,7 +320,7 @@ int pa_cury(FILE* f);
 int pa_curbnd(FILE* f);
 void pa_select(FILE* f, int u, int d);
 void pa_event(FILE* f, pa_evtrec* er);
-void pa_timer(FILE* f, int i, int t, int r);
+void pa_timer(FILE* f, int i, long t, int r);
 void pa_killtimer(FILE* f, int i);
 int pa_mouse(FILE* f);
 int pa_mousebutton(FILE* f, int m);
